Fixed names, descriptions, and removed empty 341 scenario
