+++ benchmarking/evaluations/scenarios/cosc341/concentrate_timeslot_diversify_gender_min_2_and_diversify_year_level.py
@@ -1,28 +1,30 @@
 from typing import List
 
 from api.models.enums import (
-    Gender,
     DiversifyType,
     ScenarioAttribute,
+    Gender,
     TokenizationConstraintDirection,
 )
 from api.models.tokenization_constraint import TokenizationConstraint
-from benchmarking.evaluations.goals import DiversityGoal, WeightGoal
+from benchmarking.evaluations.goals import WeightGoal, DiversityGoal
 from benchmarking.evaluations.interfaces import Scenario, Goal
 
 
-class ConcentrateTimeslotConcentrateGenderDiversifyYearLevel(Scenario):
-    # This scenario concentrates timeslot, concentrates gender, and diversifies year level.
+class ConcentrateTimeSlotDiversifyGenderMin2AndDiversifyYearLevel(Scenario):
+    # This Scenario concentrates on timeslot, diversifies female gender with min 2, and diversifies year level.
     def __init__(
-            self,
-            max_num_choices: int,
+        self,
+        max_num_choices: int,
+        value_of_female: int = Gender.FEMALE,
     ):
         super().__init__()
+        self.value_of_female = value_of_female
         self.max_num_choices = max_num_choices
 
     @property
     def name(self):
-        return "Concentrate on Timeslot, Diversify All Genders Min of Two, and Diversify Year Level"
+        return "Concentrate on Timeslot, Diversify Female Min of Two, and Diversify Year Level"
 
     @property
     def goals(self) -> List[Goal]:
@@ -33,39 +35,47 @@ class ConcentrateTimeslotConcentrateGenderDiversifyYearLevel(Scenario):
                 max_num_choices=self.max_num_choices,
             ),
             DiversityGoal(
-                DiversifyType.CONCENTRATE,
+                DiversifyType.DIVERSIFY,
                 ScenarioAttribute.GENDER.value,
-                max_num_choices=1
+                tokenization_constraint=TokenizationConstraint(
+                    direction=TokenizationConstraintDirection.MIN_OF,
+                    threshold=2,
+                    value=self.value_of_female,
+                ),
             ),
             DiversityGoal(
                 DiversifyType.DIVERSIFY,
                 ScenarioAttribute.YEAR_LEVEL.value,
-                max_num_choices=1
             ),
             WeightGoal(diversity_goal_weight=1),
         ]
 
 
-class ConcentrateGenderConcentrateTimeslotDiversifyYearLevel(Scenario):
-    # This scenario  concentrates gender, concentrates timeslot, and diversifies year level.
+class DiversifyGenderMin2ConcentrateTimeSlotAndDiversifyYearLevel(Scenario):
     def __init__(
-            self,
-            max_num_choices: int,
+        self,
+        max_num_choices: int,
+        value_of_female: int = Gender.FEMALE,
     ):
         super().__init__()
+        self.value_of_female = value_of_female
         self.max_num_choices = max_num_choices
 
     @property
     def name(self):
-        return "Concentrate on Timeslot, Diversify All Genders Min of Two, and Diversify Year Level"
+        return "Diversify Female Min of Two, Concentrate Timeslot, and Diversify Year Level"
 
     @property
     def goals(self) -> List[Goal]:
         return [
             DiversityGoal(
-                DiversifyType.CONCENTRATE,
+                DiversifyType.DIVERSIFY,
                 ScenarioAttribute.GENDER.value,
-                max_num_choices=1
+                tokenization_constraint=TokenizationConstraint(
+                    direction=TokenizationConstraintDirection.MIN_OF,
+                    threshold=2,
+                    value=self.value_of_female,
+                ),
             ),
             DiversityGoal(
                 DiversifyType.CONCENTRATE,
@@ -75,7 +85,6 @@ class ConcentrateGenderConcentrateTimeslotDiversifyYearLevel(Scenario):
             DiversityGoal(
                 DiversifyType.DIVERSIFY,
                 ScenarioAttribute.YEAR_LEVEL.value,
-                max_num_choices=1
             ),
             WeightGoal(diversity_goal_weight=1),
         ]
